Fix deadlock in FrontendHTTPReader
In the situation in which a client disconnects before all data from
the backend has been sent to the client a deadlock can occur.  In this
situation the data channel is still being written too but there is no
consumer of the channel causing the channel write to block forever.  The
data channel is only purge on Close, but the blocked write to the data
channel will prevent Close from happening.